+++ frontend/src/container/TreeListScreen/TreeListScreen.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import SearchScreen from '../SearchScreen/SearchScreen'
+
+export default function TreeListScreen() {
+    return (
+        <div style={{ padding: '10px' }}>
+            <SearchScreen />
+            <div>tree list 들어 갈곳.</div>
+        </div>
+    )
+}
